+++ packages/maas/frontend/src/app/hooks/useGovernanceAffectedModels.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { useMaaSGovernanceContext } from '~/app/context/MaaSGovernanceContext';
-import type { AffectedModel } from '~/app/shared/Phase/AffectedModelsTable';
+import type { AffectedModel } from '~/app/types/maas-model';
 import type { MaaSAuthPolicy, MaaSSubscription } from '~/app/types/subscriptions';
 import {
   getAffectedModelsFromRefs,
@@ -15,27 +15,27 @@ const isDegradedPhase = (phase: string | undefined): boolean =>
 export const useSubscriptionAffectedModels = (
   subscription: Pick<MaaSSubscription, 'modelRefs' | 'phase'>,
 ): AffectedModel[] | undefined => {
-  const { modelRefs, overviewLoaded } = useMaaSGovernanceContext();
+  const { modelRefs, loaded } = useMaaSGovernanceContext();
 
   return React.useMemo(() => {
-    if (!overviewLoaded || !isDegradedPhase(subscription.phase)) {
+    if (!loaded || !isDegradedPhase(subscription.phase)) {
       return undefined;
     }
     return getAffectedModelsFromRefs(subscription.modelRefs, modelRefs);
-  }, [overviewLoaded, subscription.modelRefs, subscription.phase, modelRefs]);
+  }, [loaded, subscription.modelRefs, subscription.phase, modelRefs]);
 };
 
 export const usePolicyAffectedModels = (
   policy: Pick<MaaSAuthPolicy, 'modelRefs' | 'phase'>,
 ): AffectedModel[] | undefined => {
-  const { modelRefs, overviewLoaded } = useMaaSGovernanceContext();
+  const { modelRefs, loaded } = useMaaSGovernanceContext();
 
   return React.useMemo(() => {
-    if (!overviewLoaded || !isDegradedPhase(policy.phase)) {
+    if (!loaded || !isDegradedPhase(policy.phase)) {
       return undefined;
     }
     return getAffectedModelsFromRefs(policy.modelRefs, modelRefs);
-  }, [overviewLoaded, policy.modelRefs, policy.phase, modelRefs]);
+  }, [loaded, policy.modelRefs, policy.phase, modelRefs]);
 };
 
 /** Resolve affected models by resource name when only overview summary fields are available. */
@@ -44,10 +44,10 @@ export const useGovernanceResourceAffectedModels = (
   phase: string | undefined,
   resourceType: PhaseResourceType,
 ): AffectedModel[] | undefined => {
-  const { subscriptions, policies, modelRefs, overviewLoaded } = useMaaSGovernanceContext();
+  const { subscriptions, policies, modelRefs, loaded } = useMaaSGovernanceContext();
 
   return React.useMemo(() => {
-    if (!overviewLoaded || !isDegradedPhase(phase)) {
+    if (!loaded || !isDegradedPhase(phase)) {
       return undefined;
     }
     if (resourceType === PhaseResourceType.SUBSCRIPTION) {
@@ -61,5 +61,5 @@ export const useGovernanceResourceAffectedModels = (
       return policy ? getAffectedModelsFromRefs(policy.modelRefs, modelRefs) : undefined;
     }
     return undefined;
-  }, [overviewLoaded, name, phase, resourceType, subscriptions, policies, modelRefs]);
+  }, [loaded, name, phase, resourceType, subscriptions, policies, modelRefs]);
 };
